Foreign.CUDA.Solver.Dense.Context

Description

Context management

newtype Handle

An opaque handle to the cuSolverDN context, which is passed to all library function calls.

http://docs.nvidia.com/cuda/cusolver/index.html#cuSolverDNhandle

Constructors

Handle

Fields

useHandle :: Ptr ()

create :: IO Handle

This function initializes the cuSolverDN library and creates a handle to the cuSolverDN context. It must be called before any other cuSolverDN API function is invoked. It allocates hardware resources necessary for accessing the GPU.

http://docs.nvidia.com/cuda/cusolver/index.html#cuSolverDNcreate

destroy :: Handle -> IO ()

This function releases resources used by the cuSolverDN library.

http://docs.nvidia.com/cuda/cusolver/index.html#cuSolverDNdestroy