Foreign.CUDA.Solver.Error

Description

Documentation

data Status

Error codes used by cuSolver library functions

http://docs.nvidia.com/cuda/cusolver/index.html#cuSolverSPstatus

Constructors

Success

NotInitialized

AllocFailed

InvalidValue

ArchMismatch

MappingError

ExecutionFailed

InternalError

MatrixTypeNotSupported

NotSupported

ZeroPivot

InvalidLicense

describe :: Status -> String

data CUSolverException

Constructors

ExitCode Status

UserError String

cusolverError :: String -> IO a

Raise a CUSolverException in the IO Monad

requireSDK :: Name -> Double -> IO a

A specially formatted error message

resultIfOk :: (Status, a) -> IO a

Return the results of a function on successful execution, otherwise throw an exception with an error string associated with the return code

nothingIfOk :: Status -> IO ()

Throw an exception with an error string associated with an unsuccessful return code, otherwise return unit.

checkStatus :: CInt -> IO ()

Throw an error if given error code is not CUSPARSE_STATUS_SUCCESS