Foreign.CUDA.Solver.Sparse.Context

Description

Context management

newtype Handle

An opaque handle to the cuSolverSP context, which is passed to all library function calls.

http://docs.nvidia.com/cuda/cusolver/index.html#cuSolverSPhandle

Constructors

Handle

Fields

useHandle :: Ptr ()

create :: IO Handle

This function initializes the cuSolverSP library and creates a handle to the cuSolverSP context. It must be called before any other cuSolverSP API function is invoked. It allocates hardware resources necessary for accessing the GPU.

http://docs.nvidia.com/cuda/cusolver/index.html#cusolverecreate

destroy :: Handle -> IO ()

This function releases resources used by the cuSolverSP library.

http://docs.nvidia.com/cuda/cusolver/index.html#cusparsedestroy