Foreign.CUDA.Solver.Sparse.High

Description

For more information see the cuSolver function reference:

http://docs.nvidia.com/cuda/cusolver/index.html#cusolver-high-level-function-reference

Documentation

data Handle

An opaque handle to the cuSolverSP context, which is passed to all library function calls.

http://docs.nvidia.com/cuda/cusolver/index.html#cuSolverSPhandle

data MatrixDescriptor :: *

An opaque type used to describe the shape and properties of a matrix.

http://docs.nvidia.com/cuda/cusparse/index.html#cusparsematdescrt

scsrlsvqr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Float -> Float -> Int -> DevicePtr Float -> IO Int

dcsrlsvqr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Double -> Double -> Int -> DevicePtr Double -> IO Int

ccsrlsvqr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Float) -> Float -> Int -> DevicePtr (Complex Float) -> IO Int

zcsrlsvqr :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Double) -> Double -> Int -> DevicePtr (Complex Double) -> IO Int

scsrlsvchol :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Float -> Float -> Int -> DevicePtr Float -> IO Int

dcsrlsvchol :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr Double -> Double -> Int -> DevicePtr Double -> IO Int

ccsrlsvchol :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Float) -> Float -> Int -> DevicePtr (Complex Float) -> IO Int

zcsrlsvchol :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> DevicePtr (Complex Double) -> Double -> Int -> DevicePtr (Complex Double) -> IO Int

scsreigvsi :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Float -> DevicePtr Int32 -> DevicePtr Int32 -> Float -> DevicePtr Float -> Int -> Float -> DevicePtr Float -> DevicePtr Float -> IO ()

dcsreigvsi :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr Double -> DevicePtr Int32 -> DevicePtr Int32 -> Double -> DevicePtr Double -> Int -> Double -> DevicePtr Double -> DevicePtr Double -> IO ()

ccsreigvsi :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Float) -> DevicePtr Int32 -> DevicePtr Int32 -> Complex Float -> DevicePtr (Complex Float) -> Int -> Float -> DevicePtr (Complex Float) -> DevicePtr (Complex Float) -> IO ()

zcsreigvsi :: Handle -> Int -> Int -> MatrixDescriptor -> DevicePtr (Complex Double) -> DevicePtr Int32 -> DevicePtr Int32 -> Complex Double -> DevicePtr (Complex Double) -> Int -> Double -> DevicePtr (Complex Double) -> DevicePtr (Complex Double) -> IO ()